requirements.txt
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
jupyter==1.0.0
openpyxl==3.1.2
joblib==1.2.0
pyyaml==6.0
imbalanced-learn==0.10.1
xgboost==1.7.5